REF 7 & 8) Moved connect method and townSocketAdapter to controller file.
